fix(pools): refactor packages
